refactor(examples): standardize imports and improve type annotations
- Update example and replay bots to use balatrobot imports
- Add proper return type annotation to connect() method
- Improve send_message() parameter handling with None default
